refactor: move char limit confirmation to shared useTranslationCell hook
Extract isOverCharLimit and handleSaveWithConfirmation from
TranslationsList/TranslationWrite into useTranslationCell so both
table and list views share the same logic. This also fixes the
missing char limit counter in the table view.
